order_microservice/requirements.txt
fastapi==0.115.6
psycopg2-binary==2.9.10
kafka-python-ng==2.2.3
uvicorn==0.34.0
alembic==1.14.0